requirements.txt
asn1crypto==1.5.1
beautifulsoup4==4.12.3
boto3==1.34.69
botocore==1.34.69
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
jmespath==1.0.1
lxml==5.1.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redshift-connector==2.1.0
requests==2.31.0
requests-oauthlib==1.4.1
s3transfer==0.10.1
scramp==1.4.4
six==1.16.0
soupsieve==2.5
tweepy==4.14.0
tzdata==2024.1
urllib3==1.26.18
